Fix the GraphQL RunOnVirtualThreadIT in native
