+++ integration-tests/virtual-threads/graphql-virtual-threads/src/test/java/io/quarkus/virtual/graphql/AbstractGraphQLTest.java
@@ -16,8 +16,6 @@ import org.hamcrest.CoreMatchers;
 
 import io.restassured.RestAssured;
 import io.restassured.parsing.Parser;
-import io.vertx.core.Context;
-import io.vertx.core.Vertx;
 
 /**
  * Some shared methods
@@ -112,72 +110,4 @@ public abstract class AbstractGraphQLTest {
         PROPERTIES.put("smallrye.graphql.events.enabled", "true");
     }
 
-    /**
-     * Hold info about a thread
-     */
-    public static class TestThread {
-
-        private long id;
-        private String name;
-        private int priority;
-        private String state;
-        private String group;
-
-        public TestThread() {
-            super();
-        }
-
-        public TestThread(long id, String name, int priority, String state, String group) {
-            this.id = id;
-            this.name = name;
-            this.priority = priority;
-            this.state = state;
-            this.group = group;
-        }
-
-        public long getId() {
-            return id;
-        }
-
-        public void setId(long id) {
-            this.id = id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public int getPriority() {
-            return priority;
-        }
-
-        public void setPriority(int priority) {
-            this.priority = priority;
-        }
-
-        public String getState() {
-            return state;
-        }
-
-        public void setState(String state) {
-            this.state = state;
-        }
-
-        public String getGroup() {
-            return group;
-        }
-
-        public void setGroup(String group) {
-            this.group = group;
-        }
-
-        public String getVertxContextClassName() {
-            Context vc = Vertx.currentContext();
-            return vc.getClass().getName();
-        }
-    }
 }
