+++ integration-tests/virtual-threads/graphql-virtual-threads/src/test/java/io/quarkus/virtual/graphql/RunOnVirtualThreadTest.java
@@ -1,8 +1,5 @@
 package io.quarkus.virtual.graphql;
 
-import org.eclipse.microprofile.graphql.GraphQLApi;
-import org.eclipse.microprofile.graphql.Mutation;
-import org.eclipse.microprofile.graphql.Query;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
@@ -11,7 +8,6 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit5.virtual.ShouldNotPin;
 import io.quarkus.test.junit5.virtual.VirtualThreadUnit;
 import io.restassured.RestAssured;
-import io.smallrye.common.annotation.RunOnVirtualThread;
 
 @QuarkusTest
 @VirtualThreadUnit
@@ -103,53 +99,4 @@ class RunOnVirtualThreadTest extends AbstractGraphQLTest {
                         Matchers.equalTo("io.vertx.core.impl.DuplicatedContext"));
     }
 
-    @GraphQLApi
-    public static class RunOnVirtualThreadObjectTestThreadResource {
-
-        // Return type Object with @RunOnVirtualThread
-        @Query
-        @RunOnVirtualThread
-        public TestThread annotatedRunOnVirtualThreadObject() {
-            sleep();
-            return getTestThread();
-        }
-
-        // Return type Object with @RunOnVirtualThread
-        @Mutation
-        @RunOnVirtualThread
-        public TestThread annotatedRunOnVirtualThreadMutationObject(String test) {
-            sleep();
-            return getTestThread();
-        }
-
-        @Query
-        @RunOnVirtualThread
-        public TestThread pinThread() {
-            // Synchronize on an object to cause thread pinning
-            Object lock = new Object();
-            synchronized (lock) {
-                sleep();
-            }
-            return getTestThread();
-        }
-
-        private void sleep() {
-            try {
-                Thread.sleep(100);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new RuntimeException(e);
-            }
-        }
-
-        private TestThread getTestThread() {
-            Thread t = Thread.currentThread();
-            long id = t.getId();
-            String name = t.getName();
-            int priority = t.getPriority();
-            String state = t.getState().name();
-            String group = t.getThreadGroup().getName();
-            return new TestThread(id, name, priority, state, group);
-        }
-    }
 }
